Refactor json-dependencies-loader to use plugins
This makes it rather more flexible, allowing it to be used for more
use cases.

One such example is resolving a reference in more than one way, in
order to both copy referenced data to the output dir, and use it in a
loader. This can be achived by applying the json-dependencies-loader
twice with different options.