Implement loading item links differently based on role
The ItemDataLinkDependencyPlugin — used with json-dependencies-loader —
can create module requests for linked resources, with varying request
query values, allowing dependencies to be loaded multiple times with
different loaders. E.g. one loader might extract a resource to the
output, while another loads its data to merge with the item (think
linked TEI data which is both the basis for an item, and needs to
be available downstream).